tests/wpt/tests/fullscreen/api/document-fullscreen-enabled.html

<!DOCTYPE html>
<title>Document#fullscreenEnabled</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe></iframe>
<script>
test(function()
{
    assert_true(document.fullscreenEnabled, "top-level document");
    const iframe = document.querySelector("iframe");
    assert_true(iframe.contentDocument.fullscreenEnabled, "same-origin iframe");
});

// The cross-origin iframe case is tested in
// /html/semantics/embedded-content/the-iframe-element/iframe-allowfullscreen.html
</script>
